+++ NEWS.md
@@ -1,3 +1,7 @@
 # 0.1
-- 0.1.4: renamed the package to `tidycensuskr.sf`
+## 0.1.4
+- renamed the package to `tidycensuskr.sf`
+- Added `convert_basic_adm2()` to convert basic adm2 data to basic local government boundaries.
+
+## 0.1.2
 - Dropped `load_districts()`